+++ RELEASE-NOTES.md
@@ -55,6 +55,7 @@ Project Page: https://github.com/a2stuff/a2d
 * Reduce cursor flickering when copying / deleting files.
 * Reduce cursor flickering while list view rows are drawn.
 * Don't repaint scrollbars before window when resizing.
+* Eliminate scrollbar repaint after window opens.
 
 ### Selector